refactor(service-common): move src files to a src folder
