+++ services/clamav/package.json
@@ -2,7 +2,7 @@
   "name": "pins-clamav",
   "version": "1.0.0",
   "description": "common service for clam antivirus",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "lint": "eslint ./",
     "lint:fix": "eslint ./ --fix",
