+++ services/mime-validation/package.json
@@ -2,7 +2,7 @@
   "name": "pins-mime-validation",
   "version": "1.0.0",
   "description": "validates mime types",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "lint": "eslint ./",
     "lint:fix": "eslint ./ --fix",
