+++ services/mime-validation/src/file_example.jpeg
