+++ services/mime-validation/src/file_example.ods
