+++ services/notify/package.json
@@ -1,8 +1,8 @@
 {
   "name": "pins-notify",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "description": "common notify service ",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "lint": "eslint ./",
     "lint:fix": "eslint ./ --fix",
