+++ services/notify/src/config.js
