+++ services/views/package.json
@@ -2,15 +2,14 @@
   "name": "pins-views",
   "version": "1.0.0",
   "description": "nunjucks views used by pins",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "lint": "eslint ./",
     "lint:fix": "eslint ./ --fix",
     "format": "prettier --write './**/*.js'",
     "test": "jest --coverage --maxWorkers=4"
   },
-  "dependencies": {
-  },
+  "dependencies": {},
   "devDependencies": {
     "eslint": "^7.11.0",
     "eslint-config-airbnb-base": "^14.2.0",
